Add support for a multi-StompletContainer message-conduit factory and server.
Each vhost is a handled by a discrete StompletContainer.  The same StompletContainer
can be used for multiple host-names through multiple registrations.  Distinct instances
are not required.

Additionally, support for a default container for clients not
caring to supply a host: header.  By default, the first registered virtual host is
considered the default, but this may be explicitly set using setDefaultContainer(...)
on the factory or the server.

Currently the client does not support vhosting; that's next.